faet(appliance): image.repository and image.tag
Can be used insetad of `source`, so that appliance-operator is able to deploy the appliance with a custom applier(other than the default `helmfile-applier`.

You can use a custom applier for air-gapped deployments for example. That is, bundle all the deployment assets into a container image so that your applier is able to deploy the appliance without an external source like Git, S3, and so on.